更新：将更新租户数据库连接的方法更换位置
