+++ modules.go
@@ -7,4 +7,5 @@ import (
 	_ "github.com/lm1996-mojor/go-core-library/log"
 	_ "github.com/lm1996-mojor/go-core-library/middleware"
 	_ "github.com/lm1996-mojor/go-core-library/redis"
+	_ "github.com/lm1996-mojor/go-core-library/tasker_factory"
 )
